Restrict sensitive models and split observability tests
